🐛 fix build error when using as dependency (#46)
See #45 for context of how this was found and how it can be reproduced. Fixes #45.

This fixes Webpack module resolution configuration so that it works correctly in all sensei use cases.

If also improves error handling in the CLI. Errors thrown from the Webpack build or child processes are properly reported, and unhandled promise rejections trigger a non-zero exit code.