♻ merge pdf package in main package (#121)
This simplifies setup and avoids a recuring issue when running "npm install" in the main package would result in infinite looping of calling the postinstall script. It also makes sensei publishable to npm more easily.

The PDF part was originally split for the main package to avoid having to install Chrome twice when developing, often in cases where I wasn't even touching the PDF part. However, as mentioned by @patou, a better technique to do that is to run `npm install` with `PUPPETEER_SKIP_CHROMIUM_DOWNLOAD` defined.

This also downgrades puppeteer from 11.0.0 to 10.4.0 to align to the version used by decktape 3.3.0 and allow npm to dedupe puppeteer and reduce the number of Chrome installs to 1. It does work in contrary from what I said in #120.

Due to https://github.com/puppeteer/puppeteer/issues/6586#issuecomment-958379915, chrome was not being downloaded when running npm install. I had to fork and patch decktape. See https://github.com/astefanutti/decktape/pull/237. The patch has been merged but not release so for now package.json points to the master branch of decktape instead of release. This should be changed as soon as decktape has a new release.