♻ migrate to webpack asset modules
Historically we used html-loader (along with extract-loader) to process the HTML templates in order to have their dependencies processed (eg images). I could not make this system work with Webpack 5 asset modules (maybe because extract-loader is not maintained anymore and can't handle asset modules), so I found an alternative [here](https://github.com/jantimon/html-webpack-plugin/issues/1557#issuecomment-908992268), which is too use `<%= require(...) %>` inside the template to have the dependencies processed directly through HtmlWebpackPlugin's loader. This simplifies the Webpack config quite alot for the small price of the special syntax, so I think it's worth it.