⚡ insert content at build time (#125)
Instead of embedding the content (including title and version) of slides and workbook in JS bundles that insert it in the DOM at run-time, make webpack do the insertion at build-time. Pros: HTML is immediately available to the browser so the pages should be faster, and index.js doesn't have to contain the full slide content just to get the title. Cons: I can't see any?

## How it works

I use 'require' inside of the HTML templates to import titles, content, and versions. This has the effect of inserting what's exported by the required module as plain text in the HTML. In order to get the content I could already do `require('slides.json').join('')`. For the title and version it was not possible as-is so I modified the revealjs loader so that 1) it could insert the version itself (passed through a loader option) 2) it could extract the title from the markdown (when signaled to do so through a loader query parameter). I then reworked the slides-json loader so that it exports both the content and the title. From there it was easy to make those available to the HTML.