🐛 fix no fonts on workbook pdf (#130)
Fonts sometimes do not load when loading the workbook in puppeteer from a file.

To fix this and also remove other inconsistencies when loading from a file vs a server, I think we should always use a server. So now, when generating PDFs, sensei runs a server that servers production build files, then run puppeteer against that server.